+++ src/main/java/nextstep/member/datamapper/MemberMapper.java
@@ -1,7 +1,6 @@
 package nextstep.member.datamapper;
 
 import nextstep.member.domain.Member;
-import nextstep.member.domain.MemberWithId;
 import nextstep.member.dto.MemberResponse;
 import nextstep.member.entity.MemberEntity;
 import org.mapstruct.Mapper;
@@ -13,8 +12,6 @@ public interface MemberMapper {
     MemberMapper INSTANCE = Mappers.getMapper(MemberMapper.class);
 
     Member entityToDomain(MemberEntity memberEntity);
-
-    MemberWithId entityToDomainWithId(MemberEntity memberEntity);
 
     default MemberEntity domainToEntity(Member member) {
         if (member == null) {
@@ -30,5 +27,5 @@ public interface MemberMapper {
         );
     }
 
-    MemberResponse domainToResponseDto(Member member);
+    MemberResponse entityToResponseDto(MemberEntity memberEntity);
 }
