+++ src/main/java/nextstep/member/domain/MemberForAuth.java
@@ -2,7 +2,8 @@ package nextstep.member.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import nextstep.auth.util.PasswordUtil;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
 @Getter
 @AllArgsConstructor
@@ -18,5 +19,23 @@ public class MemberForAuth {
 
     public final void encryptPassword() {
         this.password = PasswordUtil.encrypt(this.password);
+    }
+
+    private static final class PasswordUtil {
+
+        private static final PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+
+        private PasswordUtil() {
+        }
+
+        public static boolean matches(String rawPassword, String encryptedPassword) {
+
+            return passwordEncoder.matches(rawPassword, encryptedPassword);
+        }
+
+        public static String encrypt(String password) {
+
+            return passwordEncoder.encode(password);
+        }
     }
 }
