+++ src/main/java/nextstep/member/repository/MemberRepository.java
@@ -1,7 +1,7 @@
 package nextstep.member.repository;
 
 import nextstep.member.domain.Member;
-import nextstep.member.domain.MemberWithId;
+import nextstep.member.entity.MemberEntity;
 
 import java.util.Optional;
 
@@ -9,7 +9,7 @@ public interface MemberRepository {
 
     Long save(Member member);
 
-    Optional<Member> findById(Long id);
+    Optional<MemberEntity> findById(Long id);
 
-    Optional<MemberWithId> findByUsername(String username);
+    Optional<MemberEntity> findByUsername(String username);
 }
