+++ src/main/java/nextstep/member/repository/MemberRepositoryImpl.java
@@ -3,13 +3,11 @@ package nextstep.member.repository;
 import nextstep.member.dao.MemberDao;
 import nextstep.member.datamapper.MemberMapper;
 import nextstep.member.domain.Member;
-import nextstep.member.domain.MemberWithId;
 import nextstep.member.entity.MemberEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
-import java.util.function.Function;
 
 @Repository
 public class MemberRepositoryImpl implements MemberRepository {
@@ -28,22 +26,14 @@ public class MemberRepositoryImpl implements MemberRepository {
     }
 
     @Override
-    public Optional<Member> findById(Long id) {
-        MemberEntity memberEntity = memberDao.findById(id);
+    public Optional<MemberEntity> findById(Long id) {
 
-        return entityToOptionalDomain(memberEntity, MemberMapper.INSTANCE::entityToDomain);
+        return Optional.ofNullable(memberDao.findById(id));
     }
 
     @Override
-    public Optional<MemberWithId> findByUsername(String username) {
-        MemberEntity memberEntity = memberDao.findByUsername(username);
+    public Optional<MemberEntity> findByUsername(String username) {
 
-        return entityToOptionalDomain(memberEntity, MemberMapper.INSTANCE::entityToDomainWithId);
-    }
-
-    private <T> Optional<T> entityToOptionalDomain(MemberEntity memberEntity, Function<MemberEntity, T> function) {
-
-        return Optional.ofNullable(memberEntity)
-                .map(function);
+        return Optional.ofNullable(memberDao.findByUsername(username));
     }
 }
